Cdstp 44 restructure tel2 puml repo (#101)
* CDSTP-44 rename folder node_map_to_puml to walk_puml_graph

* CDSTP-44 rename tests folder for node_map_to_puml to walk_puml_graph

* CDSTP-44 Removed jAlergiaPipeline and put functionality into detect_loops

* CDSTP-44 Moved tests into correct structure

* CDSTP-44 rename pipelines to data_pipelines

* CDSTP-44 Fixed linting errors

* CDSTP-44 updated mypy errors file
