+++ tel2puml/utils_test.py
@@ -2,7 +2,7 @@
 """
 from typing import Literal
 
-from tel2puml.pipelines.data_creation import (
+from tel2puml.data_pipelines.data_creation import (
     generate_test_data_event_sequences_from_puml
 )
 from tel2puml.pv_to_puml import pv_to_puml_string, pv_to_puml_string_v2
