+++ tel2puml/walk_puml_graph/create_node_graph_from_event_graph.py
@@ -2,7 +2,7 @@
 
 from networkx import DiGraph
 
-from tel2puml.node_map_to_puml.node import NodeTuple, Node, SubGraphNode
+from tel2puml.walk_puml_graph.node import NodeTuple, Node, SubGraphNode
 from tel2puml.events import Event
 from tel2puml.loop_detection.loop_types import LoopEvent
 from tel2puml.tel2puml_types import PUMLEvent
