+++ tests/tel2puml/data_pipelines/test_data_ingestion.py
@@ -1,10 +1,10 @@
 """Tests for data ingestion pipeline"""
 from tel2puml.events import EventSet
-from tel2puml.pipelines.data_creation import (
+from tel2puml.data_pipelines.data_creation import (
     generate_test_data,
     generate_test_data_event_sequences_from_puml,
 )
-from tel2puml.pipelines.data_ingestion import (
+from tel2puml.data_pipelines.data_ingestion import (
     update_all_connections_from_data,
     get_graph_solutions_from_clustered_events,
     update_and_create_events_from_clustered_pvevents,
