+++ tests/tel2puml/loop_detection/conftest.py
