+++ tests/tel2puml/loop_detection/test_calculate_loop_components.py
