+++ tests/tel2puml/loop_detection/test_calculate_updated_graph.py
