+++ tests/tel2puml/loop_detection/test_detect_loops_loop_detection.py
