+++ tests/tel2puml/loop_detection/test_loop_event_methods.py
