+++ tests/tel2puml/loop_detection/test_sub_graph_of_loop.py
