+++ tests/tel2puml/puml_graph/test_graph_loop_extract.py
