+++ tests/tel2puml/test_check_puml_equiv.py
