+++ tests/tel2puml/walk_puml_graph/test_node_update.py
@@ -1,9 +1,9 @@
-"""Tests for the `node_map_to_puml.node_update` module."""
-from tel2puml.node_map_to_puml.node_update import (
+"""Tests for the `walk_puml_graph.node_update` module."""
+from tel2puml.walk_puml_graph.node_update import (
     update_nodes_with_break_points,
     update_nodes_with_break_points_from_loops,
 )
-from tel2puml.node_map_to_puml.node import Node
+from tel2puml.walk_puml_graph.node import Node
 from tel2puml.detect_loops import Loop
 from tel2puml.tel2puml_types import PUMLEvent
